Moving group membership to new framework
